# services/user/sdk/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "operationId": "UserController_get",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetUsersResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            },
            "put": {
                "operationId": "UserController_insert",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InsertUserResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "DuplicateEntryError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
        "/api/v1/user/{identity}": {
            "get": {
                "operationId": "UserController_getByIdentity",
                "parameters": [
                    {
                        "name": "identity",
                        "required": true,
                        "in": "path",
                        "description": "Identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserViewResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            },
            "delete": {
                "operationId": "UserController_delete",
                "parameters": [
                    {
                        "name": "identity",
                        "required": true,
                        "in": "path",
                        "description": "Identity",
                        "schema": {
                            "type": "string",
                            "minLength": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EntryNotFoundError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "UnexpectedError",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "oauth2": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Finance API",
        "description": "Finance API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "scopes": {
                            "openid": "openid",
                            "profile": "profile",
                            "email": "email"
                        },
                        "authorizationUrl": "https://dev-tzti25ivp5zmox2s.us.auth0.com/authorize"
                    }
                },
                "in": "header",
                "name": "Authorization",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "PaginatedResponsePage": {
                "type": "object",
                "properties": {
                    "count": {
                        "type": "number",
                        "description": "The amount available in this response",
                        "example": 10
                    },
                    "offset": {
                        "type": "number",
                        "description": "The starting index of the paginated data",
                        "example": 0
                    },
                    "total": {
                        "type": "number",
                        "description": "The total number of items available",
                        "example": 10
                    }
                },
                "required": [
                    "count",
                    "offset",
                    "total"
                ]
            },
            "StockDataResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "symbol": {
                        "type": "string"
                    },
                    "exchange": {
                        "type": "string"
                    },
                    "assetKind": {
                        "type": "string",
                        "enum": [
                            "CS",
                            "ADRC",
                            "ADRP",
                            "ADRR",
                            "UNIT",
                            "RIGHT",
                            "PFD",
                            "FUND",
                            "SP",
                            "WARRANT",
                            "INDEX",
                            "ETF",
                            "ETN",
                            "OS",
                            "GDR",
                            "OTHER",
                            "NYRS",
                            "AGEN",
                            "EQLK",
                            "BOND",
                            "ADRW",
                            "BASKET",
                            "LT"
                        ]
                    }
                },
                "required": [
                    "identity",
                    "symbol",
                    "exchange",
                    "assetKind"
                ]
            },
...
            "HttpErrorResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "integer",
                        "description": "The http code returned by the server"
                    },
                    "message": {
                        "type": "string",
                        "description": "A message describing the error that occurred"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
...
            "UserResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    }
                },
                "required": [
                    "identity",
                    "dateOfBirth",
                    "firstName",
                    "lastName"
                ]
            },
            "PaginatedUserResponse": {
                "type": "object",
                "properties": {
                    "page": {
                        "$ref": "#/components/schemas/PaginatedResponsePage"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResponse"
                        }
                    }
                },
                "required": [
                    "page",
                    "data"
                ]
            },
            "GetUsersResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PaginatedUserResponse"
                    }
                },
                "required": [
                    "success",
                    "data"
                ]
            },
            "StockOrderResponse": {
                "type": "object",
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "usdPrice": {
                        "type": "number"
                    }
                },
                "required": [
                    "amount",
                    "usdPrice"
                ]
            },
            "StockAssetResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "orders": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StockOrderResponse"
                        }
                    },
                    "stockData": {
                        "$ref": "#/components/schemas/StockDataResponse"
                    }
                },
                "required": [
                    "identity",
                    "orders",
                    "stockData"
                ]
            },
            "RealEstateAssetResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    }
                },
                "required": [
                    "identity",
                    "address"
                ]
            },
            "AssetResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "stockAsset": {
                        "nullable": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/StockAssetResponse"
                            }
                        ]
                    },
                    "realEstateAsset": {
                        "nullable": true,
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RealEstateAssetResponse"
                            }
                        ]
                    }
                },
                "required": [
                    "identity",
                    "stockAsset",
                    "realEstateAsset"
                ]
            },
            "AssetGroupResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "assets": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AssetResponse"
                        }
                    }
                },
                "required": [
                    "identity",
                    "name",
                    "assets"
                ]
            },
            "BalanceResponse": {
                "type": "object",
                "properties": {
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    }
                },
                "required": [
                    "amount",
                    "currency"
                ]
            },
            "BankAccountResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "bank": {
                        "type": "string"
                    },
                    "balance": {
                        "$ref": "#/components/schemas/BalanceResponse"
                    }
                },
                "required": [
                    "identity",
                    "bank",
                    "balance"
                ]
            },
            "ActiveIncomeResponse": {
                "type": "object",
                "properties": {
                    "monthlySalary": {
                        "type": "number"
                    }
                },
                "required": [
                    "monthlySalary"
                ]
            },
            "JobResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "activeIncome": {
                        "$ref": "#/components/schemas/ActiveIncomeResponse"
                    }
                },
                "required": [
                    "identity",
                    "title",
                    "activeIncome"
                ]
            },
            "UserViewResponse": {
                "type": "object",
                "properties": {
                    "identity": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "assetGroups": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AssetGroupResponse"
                        }
                    },
                    "bankAccounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BankAccountResponse"
                        }
                    },
                    "jobs": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/JobResponse"
                        }
                    }
                },
                "required": [
                    "identity",
                    "dateOfBirth",
                    "firstName",
                    "lastName",
                    "assetGroups",
                    "bankAccounts",
                    "jobs"
                ]
            },
            "CreateUserBody": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string"
                    }
                },
                "required": [
                    "firstName",
                    "lastName",
                    "dateOfBirth"
                ]
            },
            "InsertUserResponse": {
                "type": "object",
                "properties": {
                    "userIdentity": {
                        "type": "string"
                    }
                },
                "required": [
                    "userIdentity"
                ]
            },
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "success"
                ]
            },
...
        }
    }
}